Move NodeInfo to components
